chore: change from yarn to pnpm
This changes the package manager for the dashboard dependencies over
for the reason that pnpm is a bit more uniformly supported and has
more features. pnpm also takes up less storage and is faster at
installing.

We can also scrap this idea and go with installing and using bun.sh
instead of pnpm